+++ lib/box2d/box2d_component.dart
@@ -3,6 +3,8 @@ import 'dart:ui';
 import 'package:box2d_flame/box2d.dart' hide Timer;
 import 'package:flame/box2d/viewport.dart';
 import 'package:flame/components/component.dart';
+import 'package:ordered_set/comparing.dart';
+import 'package:ordered_set/ordered_set.dart';
 
 abstract class Box2DComponent extends Component {
   static const int DEFAULT_WORLD_POOL_SIZE = 100;
@@ -16,9 +18,11 @@ abstract class Box2DComponent extends Component {
   int velocityIterations;
   int positionIterations;
 
-  World world;
-  List<BodyComponent> components = [];
+  /// The list of components to be updated and rendered by the Box2DComponent.
+  OrderedSet<BodyComponent> components =
+      OrderedSet(Comparing.on((c) => c.priority()));
   Viewport viewport;
+  World world;
 
   Box2DComponent({
     this.dimensions,